리그 등록 완료 및 리그 상세 구현중

+++ EveryFit/src/main/java/com/kh/EveryFit/dto/LeagueDto.java
@@ -14,7 +14,7 @@ public class LeagueDto {
 	private String leagueManager;
 	private String leagueTitle;
 	private String leagueDetail;
-	private double leagueEntryFee;
+	private Integer leagueEntryFee;
 	private Date leagueStart;
 	private Date leagueEnd;
 	private String leagueStatus;
